third_party/blink/web_tests/external/wpt/css/selectors/heading-expected.txt

This is a testharness.js-based test.
Found 182 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <h1 data-expect-match="1"></h1> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h2 data-expect-match="2"></h2> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h3 data-expect-match="3"></h3> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h4 data-expect-match="4"></h4> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h5 data-expect-match="5"></h5> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h6 data-expect-match="6"></h6> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h7 data-expect-match=""></h7> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h8 data-expect-match=""></h8> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h9 data-expect-match=""></h9> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h0 data-expect-match=""></h0> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1" aria-level="2"></h1> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <h1 data-expect-match="1"></h1> in section :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <hgroup data-expect-match=""></hgroup> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading
  Failed to execute 'matches' on 'Element': ':heading' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(1)
  Failed to execute 'matches' on 'Element': ':heading(1)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(2)
  Failed to execute 'matches' on 'Element': ':heading(2)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(3)
  Failed to execute 'matches' on 'Element': ':heading(3)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(4)
  Failed to execute 'matches' on 'Element': ':heading(4)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(5)
  Failed to execute 'matches' on 'Element': ':heading(5)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(6)
  Failed to execute 'matches' on 'Element': ':heading(6)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(7)
  Failed to execute 'matches' on 'Element': ':heading(7)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(8)
  Failed to execute 'matches' on 'Element': ':heading(8)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(9)
  Failed to execute 'matches' on 'Element': ':heading(9)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(0)
  Failed to execute 'matches' on 'Element': ':heading(0)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(0, 1, 2)
  Failed to execute 'matches' on 'Element': ':heading(0, 1, 2)' is not a valid selector.
[FAIL] <p role="heading" aria-level="1" data-expect-match=""></p> :heading(6, 7)
  Failed to execute 'matches' on 'Element': ':heading(6, 7)' is not a valid selector.
Harness: the test ran to completion.
